requirements.txt
pandas==1.3.4
selenium==4.1.3
python-dotenv==0.10.1
tqdm==4.63.0
python-telegram-bot
psycopg2-binary
SQLAlchemy==1.4.37
chromedriver-binary-auto
